fix(yuna-invite-wedding): Refactors routing and authentication logic
Simplifies routing logic and implements a gatekeeper page for authentication.
This change removes the rewrite rule in `next.config.js` and the associated logic in `Invitation.tsx`.
Instead, a new `yuna.tsx` page handles the authentication and redirects to the index page with appropriate parameters.
This improves the flow and maintainability of the application.